requirements.txt
backports.functools-lru-cache==1.5
boto==2.49.0
boto3==1.9.47
botocore==1.12.47
bz2file==0.98
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cycler==0.10.0
docutils==0.14
Flask==1.0.2
futures==3.2.0
gensim==3.6.0
google-api-python-client==1.7.4
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
gunicorn==19.9.0
httplib2==0.12.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==2.2.3
numpy==1.15.4
oauthlib==2.1.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
rsa==4.0
s3transfer==0.1.13
scipy==1.1.0
six==1.11.0
smart-open==1.7.1
subprocess32==3.5.3
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
